extracted getActivityCount func to a file
